comment/views.py
from .models import Comment
from .serializers import CommentSerializer
from rest_framework import generics, permissions

class CommentRetrieveUpdateDestroyView(generics.RetrieveUpdateDestroyAPIView):
    queryset = Comment.objects.all()
    serializer_class = CommentSerializer
    permission_classes = [permissions.IsAuthenticated]
